% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{af_matrix}
\alias{af_matrix}
\title{Get Allele frequencies for SNPs}
\usage{
af_matrix(variants, bfile = NULL, plink_bin = getOption("tools_plink"))
}
\arguments{
\item{variants}{List of variants (rsids)}

\item{bfile}{If this is provided then will use the API. Default = NULL}

\item{plink_bin}{If null and bfile is not null then will detect packaged plink binary for specific OS. Otherwise specify path to plink binary. Default = NULL}
}
\value{
matrix with allele frequencies
}
\description{
Get Allele frequencies for SNPs
}
