% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{af_matrix_local}
\alias{af_matrix_local}
\title{Get AF matrix using local plink binary and reference dataset}
\usage{
af_matrix_local(variants, bfile, plink_bin, with_alleles = TRUE)
}
\arguments{
\item{variants}{List of variants (rsids)}

\item{bfile}{Path to bed/bim/fam ld reference panel}

\item{plink_bin}{Specify path to plink binary. Default = NULL. See https://github.com/explodecomputer/plinkbinr for convenient access to plink binaries}
}
\value{
data frame
}
\description{
Get AF matrix using local plink binary and reference dataset
}
