% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl2.r
\name{afl2_rsid}
\alias{afl2_rsid}
\title{Look up allele frequencies and LD scores for 1000 genomes populations by rsid}
\usage{
afl2_rsid(rsid, reference = "1000g", opengwas_jwt = get_opengwas_jwt(), ...)
}
\arguments{
\item{rsid}{Vector of rsids}

\item{reference}{Default=\code{"1000g"}}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to \url{https://api.opengwas.io} to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}

\item{...}{Additional arguments passed to \code{api_query()}.}
}
\value{
Data frame containing ancestry specific LD scores and allele frequencies for each variant
}
\description{
Look up allele frequencies and LD scores for 1000 genomes populations by rsid
}
