% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{associations}
\alias{associations}
\title{<brief desc>}
\usage{
associations(rsid, id, proxies, r2, align_alleles, palindromes,
  maf_threshold)
}
\arguments{
\item{rsid}{<what param does>}

\item{id}{<what param does>}

\item{proxies}{<what param does>}

\item{r2}{<what param does>}

\item{align_alleles}{<what param does>}

\item{palindromes}{<what param does>}

\item{maf_threshold}{<what param does>}
}
\description{
<full description>
}
