% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{batch_from_id}
\alias{batch_from_id}
\title{Extract batch name from study ID}
\usage{
batch_from_id(id)
}
\arguments{
\item{id}{Array of study IDs}
}
\value{
Array of batch names
}
\description{
Extract batch name from study ID
}
