% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{batches}
\alias{batches}
\title{Get list of data batches in IEU GWAS database}
\usage{
batches(access_token = check_access_token())
}
\arguments{
\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data}
}
\value{
data frame
}
\description{
Get list of data batches in IEU GWAS database
}
