% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl2.r
\name{bv_ldsc}
\alias{bv_ldsc}
\title{Perform basic version of bivariate LD score regression}
\usage{
bv_ldsc(id1, id2, ancestry = "infer", snpinfo = NULL)
}
\arguments{
\item{id1}{ID of trait 1}

\item{id2}{ID of trait 2}

\item{ancestry}{ancestry of traits 1 and 2 (AFR, AMR, EAS, EUR, SAS) or 'infer' (default) in which case it will try to guess based on allele frequencies}
}
\value{
List of results
}
\description{
For best results please use the LDSC software. This version is a rough approximation that uses only 20k SNPs for the purposes of minimising server load.
}
