% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.r
\name{common_snpinfo}
\alias{common_snpinfo}
\title{Retrieve a list of ~20k variants that are common in all five human super populations}
\usage{
common_snpinfo()
}
\value{
Data frame
}
\description{
Data frame includes 1000 genomes metadata including sample sizes, allele frequency and LD score
}
