% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lds_scores.R
\name{convert_variantid}
\alias{convert_variantid}
\title{Converter for gwasglue2 system variant IDs}
\usage{
convert_variantid(afl2_list)
}
\arguments{
\item{afl2_list}{A \link{afl2_list} dataframe}
}
\value{
The same dataframe with the gwasglue2 variant IDs added
}
\description{
Converter for gwasglue2 system variant IDs
}
