% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapr.R
\name{create_finemapr_data}
\alias{create_finemapr_data}
\title{Generate data for analysis in various finemapping methods}
\usage{
create_finemapr_data(region, id, bfile = NULL, plink = NULL)
}
\arguments{
\item{region}{Region of the genome to extract eg 1:109317192-110317192"}

\item{id}{Array of GWAS studies to query. See \code{gwasinfo} for available studies}
}
\value{
Each id will be a list of z score data, ld matrix, and sample size
}
\description{
Uses the finemapr package https://github.com/variani/finemapr
}
