% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.r
\name{fill_n}
\alias{fill_n}
\title{Look up sample sizes when meta data is missing from associations}
\usage{
fill_n(d)
}
\arguments{
\item{d}{Output from \code{associations}}
}
\value{
Updated version of d
}
\description{
Look up sample sizes when meta data is missing from associations
}
