% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_opengwas_jwt}
\alias{get_opengwas_jwt}
\title{Retrieve OpenGWAS JSON Web Token from .Renviron file}
\usage{
get_opengwas_jwt()
}
\value{
JWT string
}
\description{
Retrieve OpenGWAS JSON Web Token from .Renviron file
}
