% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{gwasinfo}
\alias{gwasinfo}
\title{Get list of studies with available GWAS summary statistics through API}
\usage{
gwasinfo(id = NULL, opengwas_jwt = get_opengwas_jwt())
}
\arguments{
\item{id}{List of OpenGWAS IDs to retrieve. If \code{NULL} (default) retrieves all
available datasets}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to \url{https://api.opengwas.io} to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}
}
\value{
Dataframe of details for all available studies
}
\description{
Get list of studies with available GWAS summary statistics through API
}
