% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{gwasinfo_files}
\alias{gwasinfo_files}
\title{Get list of download URLs each file (.vcf.gz, .vcf.gz.tbi, _report.html) associated with the dataset.
The URLs will expire in 2 hours.
If a dataset is missing from the results,
that means the dataset doesn't exist or you don't have access to it.
If a dataset is in the results but some/all links are missing, that means the files are unavailable.}
\usage{
gwasinfo_files(id, opengwas_jwt = get_opengwas_jwt())
}
\arguments{
\item{id}{List of MR-Base IDs to retrieve.}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a JWT. Provide the JWT string here, or store in .Renviron under the keyname OPENGWAS_JWT.}
}
\value{
Dataframe of details for requested studies
}
\description{
Get list of download URLs each file (.vcf.gz, .vcf.gz.tbi, _report.html) associated with the dataset.
The URLs will expire in 2 hours.
If a dataset is missing from the results,
that means the dataset doesn't exist or you don't have access to it.
If a dataset is in the results but some/all links are missing, that means the files are unavailable.
}
