% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieugwasr-package.R
\docType{package}
\name{ieugwasr-package}
\alias{ieugwasr}
\alias{ieugwasr-package}
\title{ieugwasr: Interface to the 'OpenGWAS' Database API}
\description{
Interface to the 'OpenGWAS' database API \url{https://api.opengwas.io/api/}. Includes a wrapper to make generic calls to the API, plus convenience functions for specific queries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MRCIEU/ieugwasr}
  \item \url{https://mrcieu.github.io/ieugwasr/}
  \item Report bugs at \url{https://github.com/MRCIEU/ieugwasr/issues}
}

}
\author{
\strong{Maintainer}: Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})

Authors:
\itemize{
  \item Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})
  \item Gibran Hemani \email{g.hemani@bristol.ac.uk} (\href{https://orcid.org/0000-0003-0920-1055}{ORCID}) [copyright holder]
  \item Ben Elsworth \email{Ben.Elsworth@bristol.ac.uk} (\href{https://orcid.org/0000-0001-7328-4233}{ORCID})
  \item Rita Rasteiro \email{rita.rasteiro@bristol.ac.uk} (\href{https://orcid.org/0000-0002-4217-3060}{ORCID})
}

}
\keyword{internal}
