% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_matrix.R
\name{ld_matrix_local}
\alias{ld_matrix_local}
\title{Get LD matrix using local plink binary and reference dataset}
\usage{
ld_matrix_local(
  variants,
  bfile,
  plink_bin,
  with_alleles = TRUE,
  tmpdir = tempdir()
)
}
\arguments{
\item{variants}{List of variants (rsids)}

\item{bfile}{Path to bed/bim/fam ld reference panel}

\item{plink_bin}{Specify path to plink binary. Default = \code{NULL}.
See \url{https://github.com/MRCIEU/genetics.binaRies} for convenient access to plink binaries}

\item{with_alleles}{Whether to append the allele names to the SNP names.
Default: \code{TRUE}}

\item{tmpdir}{Directory in which to write the temporary files used by plink. Default = \code{\link[=tempdir]{tempdir()}}}
}
\value{
data frame
}
\description{
Get LD matrix using local plink binary and reference dataset
}
