% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{print.ApiStatus}
\alias{print.ApiStatus}
\title{Print API status}
\usage{
\method{print}{ApiStatus}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{api_status}}}

\item{...}{Unused, for extensibility}
}
\value{
Print out of API status
}
\description{
Print API status
}
