% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{print.GwasInfo}
\alias{print.GwasInfo}
\title{Print GWAS information}
\usage{
\method{print}{GwasInfo}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{gwasinfo}}}

\item{...}{Unused, for extensibility}
}
\value{
Print out of GWAS information
}
\description{
Print GWAS information
}
