% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{set_reset}
\alias{set_reset}
\title{Set the reset time for OpenGWAS allowance}
\usage{
set_reset(r)
}
\arguments{
\item{r}{The API response object}
}
\value{
None
}
\description{
This function sets the reset time for the OpenGWAS allowance based on the retry-after header
returned by the API response. It also displays a warning message indicating the time at which
the allowance will reset.
}
