% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_formats.R
\name{to_TwoSampleMR}
\alias{to_TwoSampleMR}
\title{Convert output from query to TwoSampleMR format}
\usage{
to_TwoSampleMR(x, type = "exposure")
}
\arguments{
\item{x}{Output from query e.g. associations, tophits, phewas}

\item{type}{"exposure" (default) or "outcome"}
}
\value{
data frame
}
\description{
Convert output from query to TwoSampleMR format
}
