% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tophits}
\alias{tophits}
\title{Obtain top hits from a GWAS dataset}
\usage{
tophits(id, pval = 5e-08, clump = 1, r2 = 0.001, kb = 5000,
  access_token = check_access_token())
}
\arguments{
\item{id}{Array of GWAS studies to query. See \code{gwasinfo} for available studies}

\item{pval}{P-value threshold. Default = 5e-8}

\item{clump}{Whether to clump (1) or not (0). Default = 1}

\item{r2}{Clumping r2 threshold. Default is very strict, 0.001}

\item{kb}{Clumping kb window. Default is very strict, 5000}

\item{access_token=check_access_token()}{<what param does>}
}
\value{
Dataframe
}
\description{
By default performs clumping on the server side.
}
