% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{user}
\alias{user}
\title{Get user details}
\usage{
user(opengwas_jwt = get_opengwas_jwt())
}
\arguments{
\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}
}
\value{
user information
}
\description{
Get user details
}
