% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variants_chrpos}
\alias{variants_chrpos}
\title{Obtain information about chr pos and surrounding region}
\usage{
variants_chrpos(chrpos, radius = 0, opengwas_jwt = get_opengwas_jwt())
}
\arguments{
\item{chrpos}{list of \verb{<chr>:<pos>} in build 37,
e.g. \code{c("3:46414943", "3:122991235")}. Also allows ranges e.g. \code{"7:105561135-105563135"}}

\item{radius}{Radius around each chrpos, default = \code{0}}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to \url{https://api.opengwas.io} to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}
}
\value{
Data frame
}
\description{
For a list of chromosome and positions, finds all variants within a given radius
}
