% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variants_to_rsid}
\alias{variants_to_rsid}
\title{Convert mixed array of rsid and chrpos to list of rsid}
\usage{
variants_to_rsid(variants)
}
\arguments{
\item{variants}{Array of variants e.g. c("rs234", "7:105561135-105563135")}
}
\value{
list of rsids
}
\description{
Convert mixed array of rsid and chrpos to list of rsid
}
