% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl2.r
\name{afl2_rsid}
\alias{afl2_rsid}
\title{Look up allele frequencies and LD scores for 1000 genomes populations by rsid}
\usage{
afl2_rsid(rsid, reference = "1000g")
}
\arguments{
\item{rsid}{Vector of rsids}

\item{reference}{Default=\code{"1000g"}}
}
\value{
data frame
}
\description{
Look up allele frequencies and LD scores for 1000 genomes populations by rsid
}
