% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_query}
\alias{api_query}
\title{Wrapper for sending queries and payloads to API}
\usage{
api_query(path, query = NULL, access_token = check_access_token())
}
\arguments{
\item{path}{Either a full query path (e.g. for get) or an endpoint (e.g. for post) queries}

\item{query}{If post query, provide a list of arguments as the payload. NULL by default}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data. By default, checks if already authenticated through \code{get_access_token} and if not then does not perform authentication}
}
\value{
Parsed json output from query, often in form of data frame
}
\description{
There are a number of different GET and POST endpoints in the GWAS database API. This is a generic way to access them
}
