% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{batches}
\alias{batches}
\title{Get list of data batches in IEU GWAS database}
\usage{
batches(access_token = check_access_token(), opengwas_jwt = get_opengwas_jwt())
}
\arguments{
\item{access_token}{Google OAuth2 access token.
Used to authenticate level of access to data. By default, checks if already}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}
}
\value{
data frame
}
\description{
Get list of data batches in IEU GWAS database
}
