% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_access_token}
\alias{check_access_token}
\title{Check if authentication has been made}
\usage{
check_access_token()
}
\value{
NULL or access_token depending on current authentication state
}
\description{
Deprecated. Use \code{get_opengwas_jwt()} instead. See https://mrcieu.github.io/ieugwasr/articles/guide.html#authentication for more information.
}
