% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{editcheck}
\alias{editcheck}
\title{Check datasets that are in process of being uploaded}
\usage{
editcheck(id, access_token = check_access_token())
}
\arguments{
\item{id}{ID}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data. By default, checks if already authenticated through \code{get_access_token} and if not then does not perform authentication}
}
\value{
Dataframe
}
\description{
Check datasets that are in process of being uploaded
}
