% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{editcheck}
\alias{editcheck}
\title{Check datasets that are in process of being uploaded}
\usage{
editcheck(id, opengwas_jwt = get_opengwas_jwt())
}
\arguments{
\item{id}{ID}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}
}
\value{
Dataframe
}
\description{
Check datasets that are in process of being uploaded
}
