% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl2.r, R/query.R
\name{fill_n}
\alias{fill_n}
\title{Look up sample sizes when meta data is missing from associations}
\usage{
fill_n(d)

fill_n(d)
}
\arguments{
\item{d}{Output from \code{associations}}
}
\value{
Updated version of d

Updated version of d
}
\description{
Look up sample sizes when meta data is missing from associations

Look up sample sizes when meta data is missing from associations
}
