% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl2.r, R/query.R
\name{fill_n}
\alias{fill_n}
\title{Look up sample sizes when meta data is missing from associations}
\usage{
fill_n(d)

fill_n(d)
}
\arguments{
\item{d}{Output from \code{\link{associations}}}
}
\value{
Data frame which is an updated version of the input but with sample sizes inferred where missing

Updated version of d
}
\description{
Look up sample sizes when meta data is missing from associations

Look up sample sizes when meta data is missing from associations
}
