% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fill_n}
\alias{fill_n}
\title{Look up sample sizes when meta data is missing from associations}
\usage{
fill_n(d, opengwas_jwt = get_opengwas_jwt())
}
\arguments{
\item{d}{Output from \code{\link{associations}}}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}
}
\value{
Updated version of d
}
\description{
Look up sample sizes when meta data is missing from associations
}
