% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.r
\name{infer_ancestry}
\alias{infer_ancestry}
\title{Infer ancestry of GWAS dataset by matching against 1000 genomes allele frequencies}
\usage{
infer_ancestry(d, snpinfo = NULL)
}
\arguments{
\item{d}{Data frame containing at least rsid and eaf columns. e.g. output from associations}
}
\value{
data frame ordered by most likely ancestry
}
\description{
Uses ~20k SNPs selected for common frequency across 5 major super populations
}
