% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_clump.R
\name{ld_clump}
\alias{ld_clump}
\title{Perform LD clumping on SNP data}
\usage{
ld_clump(dat = NULL, clump_kb = 10000, clump_r2 = 0.001,
  clump_p = 1, access_token = NULL, bfile = NULL, plink_bin = NULL)
}
\arguments{
\item{dat}{Output from \code{format_data}. Must have a SNP name column (SNP) and pval column called either "P", "pval.exposure", or "pval". If id or id.exposure is present then it will  or pval.exposure not present they will be generated}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data}

\item{clump_kb=10000}{Clumping window}

\item{clump_r2=0.001}{Clumping r2 cutoff. Note that this default value has recently changed from 0.01.}

\item{clump_p=1}{Clumping sig level for index SNPs}

\item{bfile=NULL}{If this is provided then will use the API}

\item{plink_bin=NULL}{If null and bfile is not null then will detect packaged plink binary for specific OS. Otherwise specify path to plink binary}
}
\value{
Data frame
}
\description{
Uses PLINK clumping method, where SNPs in LD within a particular window will be pruned. The SNP with the lowest p-value is retained.
}
