% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_clump.R
\name{ld_clump}
\alias{ld_clump}
\title{Perform LD clumping on SNP data}
\usage{
ld_clump(
  dat = NULL,
  clump_kb = 10000,
  clump_r2 = 0.001,
  clump_p = 0.99,
  pop = "EUR",
  access_token = NULL,
  bfile = NULL,
  plink_bin = NULL
)
}
\arguments{
\item{dat}{Dataframe. Must have a variant name column ("rsid") and pval column called "pval". If id is present then clumping will be done per unique id.}

\item{clump_kb}{Clumping kb window. Default is very strict, 10000}

\item{clump_r2}{Clumping r2 threshold. Default is very strict, 0.001}

\item{clump_p}{Clumping sig level for index variants. Default = 1 (i.e. no threshold)}

\item{pop}{Super-population to use as reference panel. Default = "EUR". Options are EUR, SAS, EAS, AFR, AMR}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data}

\item{bfile}{If this is provided then will use the API. Default = NULL}

\item{plink_bin}{If null and bfile is not null then will detect packaged plink binary for specific OS. Otherwise specify path to plink binary. Default = NULL}
}
\value{
Data frame
}
\description{
Uses PLINK clumping method, where variants in LD within a particular window will be pruned. The SNP with the lowest p-value is retained.
}
