% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_clump.R
\name{ld_clump_api}
\alias{ld_clump_api}
\title{Perform clumping on the chosen SNPs using through API}
\usage{
ld_clump_api(dat, clump_kb = 10000, clump_r2 = 0.1, clump_p,
  access_token = check_access_token())
}
\arguments{
\item{dat}{Output from \code{read_exposure_data}. Must have a SNP name column (SNP), SNP chromosome column (chr_name), SNP position column (chrom_start) and p-value column (pval.exposure)}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data#' @return Data frame of only independent SNPs}

\item{clump_kb=10000}{Clumping window}

\item{clump_r2=0.1}{Clumping r2 cutoff}

\item{clump_p=1}{Clumping sig level for index SNPs}
}
\description{
Perform clumping on the chosen SNPs using through API
}
