% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_clump.R
\name{ld_clump_local}
\alias{ld_clump_local}
\title{Wrapper for clump function using local plink binary and ld reference dataset}
\usage{
ld_clump_local(dat, clump_kb, clump_r2, clump_p, bfile, plink_bin)
}
\arguments{
\item{dat}{Output from \code{read_exposure_data}. Must have a SNP name column (SNP), SNP chromosome column (chr_name), SNP position column (chrom_start) and p-value column (pval.exposure)}

\item{clump_kb=10000}{Clumping window}

\item{clump_r2=0.1}{Clumping r2 cutoff}

\item{clump_p=1}{Clumping sig level for index SNPs}

\item{bfile=NULL}{Reference plink bed/bim/fam file set to be used for clumping}

\item{plink_bin=NULL}{If null then will detect packaged plink binary for specific OS. Otherwise specify path to plink binary}
}
\value{
data frame of clumped SNPs
}
\description{
Wrapper for clump function using local plink binary and ld reference dataset
}
