% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_matrix.R
\name{ld_matrix_local}
\alias{ld_matrix_local}
\title{Get LD matrix using local plink binary and reference dataset}
\usage{
ld_matrix_local(snps, bfile, plink_bin, with_alleles = TRUE)
}
\arguments{
\item{snps}{<what param does>}

\item{bfile}{<what param does>}

\item{plink_bin}{<what param does>}

\item{with_alleles=TRUE}{<what param does>}
}
\value{
data frame
}
\description{
Get LD matrix using local plink binary and reference dataset
}
