% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{phewas}
\alias{phewas}
\title{Perform fast phewas of a specific variants against all available GWAS datasets}
\usage{
phewas(
  variants,
  pval = 1e-05,
  batch = c(),
  opengwas_jwt = get_opengwas_jwt(),
  ...
)
}
\arguments{
\item{variants}{Array of variants e.g. \code{c("rs234", "7:105561135-105563135")}}

\item{pval}{p-value threshold. Default = \code{0.00001}}

\item{batch}{Vector of batch IDs to search across. If \code{c()} (default) then returns all batches}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to \url{https://api.opengwas.io} to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}

\item{...}{Unused, for extensibility}
}
\value{
Dataframe
}
\description{
This is faster than doing it manually through \code{\link{associations}}
}
