% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{tophits}
\alias{tophits}
\title{Obtain top hits from a GWAS dataset}
\usage{
tophits(id, preclumped = 1, pval = 5e-08, clump = 1, r2 = 0.001,
  kb = 10000, access_token = check_access_token())
}
\arguments{
\item{id}{Array of GWAS studies to query. See \code{gwasinfo} for available studies}

\item{preclumped}{By default will return preclumped hits. p-value threshold 5e-8, with r2 threshold 0.001 and kb threshold 10000, using only SNPs with MAF > 0.01 in the European samples in 1000 genomes. If preclumped = 0 then will use the other specified parameters to obtain tophits.}

\item{pval}{If preclumped = 0, use this p-value threshold. Default = 5e-8}

\item{clump}{If preclumped = 0, whether to clump (1) or not (0). Default = 1}

\item{r2}{If preclumped = 0, use this clumping r2 threshold. Default is very strict, 0.001}

\item{kb}{If preclumped = 0, use this clumping kb window. Default is very strict, 10000}

\item{access_token}{Google OAuth2 access token. Used to authenticate level of access to data. By default, checks if already authenticated through \code{get_access_token} and if not then does not perform authentication}
}
\value{
Dataframe
}
\description{
By default performs clumping on the server side.
}
