% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variants_gene}
\alias{variants_gene}
\title{Obtain variants around a gene}
\usage{
variants_gene(gene, radius = 0, opengwas_jwt = get_opengwas_jwt(), ...)
}
\arguments{
\item{gene}{Vector of genes, either Ensembl or Entrez,
e.g. \code{c("ENSG00000123374", "ENSG00000160791")} or \code{1017}}

\item{radius}{Radius around the gene region to include. Default = \code{0}}

\item{opengwas_jwt}{Used to authenticate protected endpoints. Login to \url{https://api.opengwas.io} to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.}

\item{...}{Additional arguments passed to \code{api_query()}.}
}
\value{
data frame with the following columns
}
\description{
Provide a gene identified, either Ensembl or Entrez
}
