% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_scores.R
\name{write_ldscores}
\alias{write_ldscores}
\title{LD scores writer
The LD scores are saved in compressed \code{.gz}files, split by chromosome name.}
\usage{
write_ldscores(afl2_list, pop = "EUR", path_to_save = pop)
}
\arguments{
\item{afl2_list}{A \link{afl2_list} dataframe.}

\item{pop}{A string with the population name. Default is "EUR".}

\item{path_to_save}{A string with the path to save the ldscores. Default is the population name.}
}
\value{
A directory with the compressed LD scores files.
}
\description{
LD scores writer
The LD scores are saved in compressed \code{.gz}files, split by chromosome name.
}
